Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/user-timing/measure_syntax_err.any-expected.txt


PASS Create a mark "existing_mark"
PASS self.performance.measure("measure", "mark"), where "mark" is a non-existent mark, throws a SyntaxError exception.
PASS self.performance.measure("measure", "mark", "existing_mark"), where "mark" is a non-existent mark, throws a SyntaxError exception.
PASS self.performance.measure("measure", "existing_mark", "mark"), where "mark" is a non-existent mark, throws a SyntaxError exception.
PASS self.performance.measure("measure", "mark", "mark"), where "mark" is a non-existent mark, throws a SyntaxError exception.
